move schemas to a separate module
